Exception Handling and Bean Validation

+++ src/main/java/com/devcaleb/ecommerce/services/ProductService.java
@@ -3,14 +3,16 @@ package com.devcaleb.ecommerce.services;
 import com.devcaleb.ecommerce.dto.ProductDTO;
 import com.devcaleb.ecommerce.entities.Product;
 import com.devcaleb.ecommerce.repositories.ProductRepository;
+import com.devcaleb.ecommerce.services.exceptions.DatabaseException;
+import com.devcaleb.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -20,10 +22,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
-        return dto;
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Resource not found!!"));
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -42,15 +42,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource not found!!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public  void delete(Long id){
-       repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not found!!!");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Referential Integrity Failure!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
